requirements.txt
streamlit
openai
python-dotenv
pyperclip
pillow
pydub
